+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 1.0.0
+
+- Upgrade `generic-pool` dependency and use `Promise`s
+- Support loading the style from a file path and implicitly set the
+[base directory](https://github.com/mapnik/node-mapnik/blob/master/src/mapnik_map.cpp#L1146) for the `mapnik.Map`
+
 ## 0.1.3
 
  - Removed use of `peerDepedencies` for mapnik (#8)